requirements.txt
pyautogui>=0.9.54
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
pynput>=1.7.6
numpy>=1.24.0
PyQt6>=6.5.0
pyinstaller>=6.0.0
